requirements.txt
flask
flask-cors
gunicorn
numpy
pandas
scikit-learn
python-dotenv
openai
google-generativeai
rapidfuzz
requests